added check-authentication.ts to re-use middleware authentication, JSDoc comments for documentation
