+++ src/misc/random-password.ts
@@ -1,5 +1,15 @@
-// create pseudo-random password for new users
+/**
+ * @file random-password.ts - function to create a pseudo-random password
+ * @author cmc
+ * @since v0.0.1
+ */
 
+/**
+ * @function create a pseudo-random password of given length
+ * @author cmc
+ * @param num The length of the password
+ * @returns The password
+ */
 export default (num: number): string => {
   const chars = 'qwertzuop+asdfghjkl#yxcvbnmQWERTZUIOP+ASDFGHJKL#YXCVBNM,.-<>1234567890!§$%&/()"\'[]{}=?'
   let pwd: string = ''
